% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_stratification_helpers.R
\name{add_pairwise_environment_metadata}
\alias{add_pairwise_environment_metadata}
\title{Add environment metadata to pairwise Lin results}
\usage{
add_pairwise_environment_metadata(pairwise, env_info)
}
\arguments{
\item{pairwise}{Pairwise result table.}

\item{env_info}{Environment metadata table.}
}
\value{
Pairwise table with metadata.
}
\description{
Add environment metadata to pairwise Lin results
}
