% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_freq.R
\name{allele_freq}
\alias{allele_freq}
\title{Computes allele frequencies for specified populations given SNP array data}
\usage{
allele_freq(geno, populations)
}
\arguments{
\item{geno}{matrix of genotypes coded as the dosage of allele B {0, 1, 2}
with individuals in rows (named) and SNPs in columns (named)}

\item{populations}{list of named populations. Each population has a vector of IDs
that belong to the population. Allele frequencies will be derived from all animals}
}
\value{
data.frame consisting of allele_frequencies for populations (columns) for
 each SNP (rows)
}
\description{
Computes allele frequencies for specified populations given SNP array data
}
