% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breed_lookup.R
\name{breed_lookup}
\alias{breed_lookup}
\title{Lookup haplotypes from a breed probability matrix}
\usage{
breed_lookup(geno, bmat, white_freq = FALSE)
}
\arguments{
\item{geno}{string containing genotypes for a region represented as (ex. 2011210?1211?)}

\item{bmat}{a breed probability matrix as described from above}

\item{white_freq}{logical. If true, combines breed probabilities of Yorkshire and Landrace breeds
when those are present.}
}
\value{
probabilities of being particular breeds based on stretch of genotype information provided
 in geno argument.
}
\description{
Breed probability matrices can be constructed with the help of haplo_comp(), they contain 
genotypes for a region in rows and probabilities that that genotype belongs to a breed / breed hybrid
in the columns.
}

