% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breedTools.R
\docType{package}
\name{breedTools}
\alias{breedTools}
\alias{breedTools-package}
\title{Estimate breed composition of unknown animals using genotyping data}
\description{
The breedTools package contains functions used to determine the breed/ancestry composition
for test animals from their genotyping data alone. It does so using genotyping data from 
a set of reference populations. The methods contained in this package were used to generate
the manuscript by Funkhouser et al. in the inaugural issue of Translational Animal Science (TAS)
from January 2016.
\cr\cr
There are a few ways in which breedTools can infer breed composition. These are:
\itemize{
   \item Predict overall breed/ancestry proportions genome wide.
   \item Estimate breed posterior probabilities based on looking at
         phased, limited regions of the genome.
}
}
