% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_local_reference.R
\name{build_local_reference}
\alias{build_local_reference}
\title{Build reference panel of breed probabilities}
\usage{
build_local_reference(chr, markers, geno_, map_, ped_ = NULL, path_ = NULL,
  groups_ = NULL, keep_fimpute_ = FALSE, parent_ = FALSE,
  reference_ = FALSE)
}
\arguments{
\item{chr}{character providing the chromosome requiring phasing}

\item{markers}{character vector containing names of probes used to construct multi-locus haplotypes}

\item{geno_}{a matrix or data.frame with genotypes of reference animals. SNPs in columns and 
individuals in rows. Genotypes must be coded as dosage of allele 'b' {0, 1, 2}.}

\item{map_}{a data.frame containing SNP map information for each SNP present in geno}

\item{ped_}{a data.frame pedigree providing family information for each individual in geno. The first
column of the pedigree is for ID, second is for sire/father ID, third is for dam/mother ID, fourth
is sex of ID {"M" or "F"}. If ids in geno are not present in pedigree, they will be added to the
end with missing parent information and a "M" sex. Correct sex information should only be required
when imputing/phasing the sex chromosomes.}

\item{path_}{a character represting the path to the FImpute binary. If omitted, assumes FImpute binary
resides along PATH.}

\item{groups_}{a list of character vectors with the names of IDs meant to be processed as groups. 
Names of list should be the names of the groups, with each element containing IDs.}

\item{keep_fimpute_}{logical if TRUE, output files generated from snpTools::fimpute_run are kept. 
By default, build_KBP will use fimpute_run to build haplotypes, but as soon as haplotypes are 
extracted, fimpute output will be deleted so not to clutter hard drive space.}

\item{parent_}{logical if TRUE, breed probabilities or haplotypes will be returned only from
parents and not from progeny,}

\item{reference_}{logical use TRUE if you want `groups` to be used to specify the reference animals
used by FImpute. Not compatable with parent TRUE.}
}
\value{
a data.frame containing breed probabilities for each genotype made possible from the
reference animal haplotypes
}
\description{
Generalized function to build a reference panel, used to estimate breed probabilities for a set 
of multi-locus genotypes from a specified region.
}
