% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_stratification_anova.R
\name{calculate_environment_stratification_anova}
\alias{calculate_environment_stratification_anova}
\title{Calculate design-aware ANOVA}
\usage{
calculate_environment_stratification_anova(data)
}
\arguments{
\item{data}{Standardized phenotype data.}
}
\value{
ANOVA table as a data frame.
}
\description{
Calculate design-aware ANOVA
}
