% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_stratification_empty.R
\name{empty_group_membership}
\alias{empty_group_membership}
\title{Empty group membership table}
\usage{
empty_group_membership()
}
\value{
Empty data frame.
}
\description{
Empty group membership table
}
