% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_stratification_empty.R
\name{empty_group_summary}
\alias{empty_group_summary}
\title{Empty group summary table}
\usage{
empty_group_summary()
}
\value{
Empty data frame.
}
\description{
Empty group summary table
}
