% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_stratification_empty.R
\name{empty_pairwise}
\alias{empty_pairwise}
\title{Empty pairwise Lin result table}
\usage{
empty_pairwise()
}
\value{
Empty data frame.
}
\description{
Empty pairwise Lin result table
}
