% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_stratification_empty.R
\name{empty_ungrouped}
\alias{empty_ungrouped}
\title{Empty ungrouped environment table}
\usage{
empty_ungrouped()
}
\value{
Empty data frame.
}
\description{
Empty ungrouped environment table
}
