% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_groups.R
\name{get_compatible_groups}
\alias{get_compatible_groups}
\title{Get compatible environment groups}
\usage{
get_compatible_groups(group_tests)
}
\arguments{
\item{group_tests}{Output from \code{\link[=make_environment_groups]{make_environment_groups()}}.}
}
\value{
A tibble containing only homogeneous groups.
}
\description{
Get compatible environment groups
}
