% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsibs.R
\name{get_gsibs}
\alias{get_gsibs}
\title{Obtain relatives of id that share a grand mother or grand father, if any relatives}
\usage{
get_gsibs(id, ped)
}
\arguments{
\item{id}{character naming ID of interest}

\item{ped}{data.frame consisting of pedigree information
with columns {ID, Sire, Dam} in that order}
}
\value{
character vector of sibling IDs
}
\description{
Obtain relatives of id that share a grand mother or grand father, if any relatives
}

