% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_groups.R
\name{get_maximal_groups}
\alias{get_maximal_groups}
\title{Select maximal compatible environment groups}
\usage{
get_maximal_groups(compatible_groups)
}
\arguments{
\item{compatible_groups}{Output from \code{\link[=get_compatible_groups]{get_compatible_groups()}}.}
}
\value{
A tibble with maximal compatible groups only.
}
\description{
Removes smaller compatible groups that are fully contained inside larger
compatible groups.
}
