% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_sibs.R
\name{get_sibs}
\alias{get_sibs}
\title{Obtain siblings of id, if any}
\usage{
get_sibs(id, ped)
}
\arguments{
\item{id}{character naming ID of interest}

\item{ped}{data.frame consisting of pedigree information
with columns {ID, Sire, Dam} in that order}
}
\value{
character vector of sibling IDs
}
\description{
Obtain siblings of id, if any
}

