% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_column.R
\name{guess_column}
\alias{guess_column}
\title{Guess a column name from possible Breedbase column names}
\usage{
guess_column(df, possible_names)
}
\arguments{
\item{df}{A data frame.}

\item{possible_names}{A character vector of possible column names.}
}
\value{
A column name if found, otherwise \code{NULL}.
}
\description{
Guess a column name from possible Breedbase column names
}
\examples{
df <- data.frame(germplasmName = "G1", phenotype = 10)
guess_column(df, c("germplasmName", "accession_name"))

}
