% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_comp.R
\name{haplo_comp}
\alias{haplo_comp}
\title{Compute haplotype frequencies and genotype probabilities}
\usage{
haplo_comp(x, y = NULL, haplotypes = FALSE)
}
\arguments{
\item{x}{A data.frame that contains haplotypes for individuals
(rows) of a population, for a number of SNPs (columns). Haplotypes coded
0 or 1 as type character.}

\item{y}{A data.frame that contains haplotypes for individuals
(rows) of a population, for a number of SNPs (columns). Haplotypes coded
0 or 1 as type character.}

\item{haplotypes}{boolean. If true, return haplotype frequencies of population x.
If false, return breed probability matrix (posterior probabilities for each breed,
rows, to have particular genotype, columns)}
}
\value{
Either haplotype frequencies or breed probability matrix
}
\description{
Calculates haplotype frequencies and uses them to estimate genotype probabilities
}

