% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_stratification_helpers.R
\name{harmonic_mean}
\alias{harmonic_mean}
\title{Calculate harmonic mean}
\usage{
harmonic_mean(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Harmonic mean.
}
\description{
Calculate harmonic mean
}
