% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hartley_fmax.R
\name{hartley_critical_sim}
\alias{hartley_critical_sim}
\title{Simulate Hartley Fmax critical value}
\usage{
hartley_critical_sim(df_error_vec, alpha = 0.05, n_sim = 1e+05, seed = 123)
}
\arguments{
\item{df_error_vec}{Numeric vector of residual degrees of freedom.}

\item{alpha}{Significance level. Default is \code{0.05}.}

\item{n_sim}{Number of simulations. Default is \code{100000}.}

\item{seed}{Random seed.}
}
\value{
A numeric critical value.
}
\description{
Estimates the Hartley Fmax critical value using simulation. This version
can handle different residual degrees of freedom across environments.
}
