% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_gsib.R
\name{has_gsib}
\alias{has_gsib}
\title{Detect if ID has a 'great sibling' in the pedigree, as specified}
\usage{
has_gsib(id, ped)
}
\arguments{
\item{id}{character naming ID of interest}

\item{ped}{data.frame consisting of pedigree information
with columns {ID, Sire, Dam} in that order}
}
\value{
boolean. TRUE if animal has a great sibling
}
\description{
'great sibling' would mean siblings related by their grandparents
}
