% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_progeny.R
\name{has_progeny}
\alias{has_progeny}
\title{Detect if each ID has progeny in pedigree and how many.}
\usage{
has_progeny(id, ped, geno = NULL)
}
\arguments{
\item{id}{character naming ID of interest}

\item{ped}{data.frame consisting of pedigree information
with columns {ID, Sire, Dam} in that order}

\item{geno}{matrix of SNP array data}
}
\value{
integer indicating how many progeny the id of interest
 has, either in the provided pedigree or in the pedgree with
 genotyping data. If none found, returns NA.
}
\description{
Genotyping data can also be provided to know whether
 an available sire has genotyping data available
}
