% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/has_sib.R
\name{has_sib}
\alias{has_sib}
\title{Detect if each ID has a full-sib or half-sib in the pedigree, as specified}
\usage{
has_sib(id, ped)
}
\arguments{
\item{id}{character naming ID of interest}

\item{ped}{data.frame consisting of pedigree information
with columns {ID, Sire, Dam} in that order}
}
\value{
boolean. TRUE if animal has a sib of specified sib_type
}
\description{
Genotyping data can also be provided to know whether
 the sib has genotyping data available
}

