% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_sire.R
\name{has_sire}
\alias{has_sire}
\title{Detect if each ID has a Father/Sire in pedigree}
\usage{
has_sire(id, ped, geno = NULL)
}
\arguments{
\item{id}{character naming ID of interest}

\item{ped}{data.frame consisting of pedigree information
with columns {ID, Sire, Dam} in that order}

\item{geno}{matrix of SNP array data}
}
\value{
boolean. TRUE if animal has a Father/Sire in the
 pedigree. If geno is provided, only returns TRUE if that
 Father/Sire has pedigree information
}
\description{
Genotyping data can also be provided to know whether
 an available sire has genotyping data available
}
