% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_environment_grouping.R
\name{lin_error_mse}
\alias{lin_error_mse}
\title{Calculate residual error MSE for a Lin environment group}
\usage{
lin_error_mse(data, envs, common_genotypes)
}
\arguments{
\item{data}{Standardized phenotype data.}

\item{envs}{Character vector of environments.}

\item{common_genotypes}{Character vector of common genotypes.}
}
\value{
A list with MSE, residual df, and message.
}
\description{
Calculate residual error MSE for a Lin environment group
}
