% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_environment_grouping.R
\name{lin_ss_ge}
\alias{lin_ss_ge}
\title{Calculate Lin genotype-by-environment sum of squares}
\usage{
lin_ss_ge(data, envs)
}
\arguments{
\item{data}{Standardized phenotype data.}

\item{envs}{Character vector of environments to test.}
}
\value{
A list with a summary table and common genotype names.
}
\description{
Calculate Lin genotype-by-environment sum of squares
}
