% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_environment_grouping.R
\name{lin_test_group}
\alias{lin_test_group}
\title{Test one Lin environment group}
\usage{
lin_test_group(data, envs, alpha = 0.05)
}
\arguments{
\item{data}{Standardized phenotype data.}

\item{envs}{Character vector of environments.}

\item{alpha}{Significance level.}
}
\value{
Data frame with Lin GE test result.
}
\description{
Test one Lin environment group
}
