% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_groups.R
\name{make_environment_groups}
\alias{make_environment_groups}
\title{Test all compatible environment groups using Hartley Fmax}
\usage{
make_environment_groups(
  env_anova,
  alpha = 0.05,
  max_group_size = Inf,
  n_sim = 1e+05,
  seed = 123
)
}
\arguments{
\item{env_anova}{A data frame from \code{\link[=analyze_environment_anova]{analyze_environment_anova()}}.}

\item{alpha}{Significance level.}

\item{max_group_size}{Maximum environment group size to test.}

\item{n_sim}{Number of simulations for Hartley critical value.}

\item{seed}{Random seed.}
}
\value{
A tibble with all tested groups.
}
\description{
Tests all combinations of environments and identifies groups with homogeneous
residual variances.
}
