% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_build.R
\name{ped_build}
\alias{ped_build}
\title{Constructs pedigree information for select individuals}
\usage{
ped_build(ids, ped, gmax = 1000)
}
\arguments{
\item{ids}{character vector containing individuals of interest}

\item{ped}{data.frame of full pedigree to subset from}

\item{gmax}{numeric max number of generations to go back}
}
\value{
data.frame of subsetted pedigree
}
\description{
Constructs pedigree information for select individuals
}

