% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_stratification_anova.R
\name{prepare_design_factors}
\alias{prepare_design_factors}
\title{Prepare design factors}
\usage{
prepare_design_factors(data)
}
\arguments{
\item{data}{Standardized phenotype data.}
}
\value{
Data frame with design columns converted to factors.
}
\description{
Prepare design factors
}
