% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented_row_column.R
\name{print.augmented_row_column_analysis}
\alias{print.augmented_row_column_analysis}
\title{Print augmented row-column analysis}
\usage{
\method{print}{augmented_row_column_analysis}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{analyze_augmented_row_column_design()}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print augmented row-column analysis
}
