% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_bp_generic.R
\name{solve_bp_generic}
\alias{solve_bp_generic}
\title{Solve generic breed probabilities}
\usage{
solve_bp_generic(snps_, geno_, b_mat_, white_)
}
\arguments{
\item{snps_}{character vector containing names of SNPs to be extracted in geno, for use
in breed probability estimation.}

\item{geno_}{a matrix or data.frame with genotypes of test animals. SNPs in columns and 
individuals in rows. Genotypes must be coded as dosage of allele 'b' {0, 1, 2}.}

\item{b_mat_}{a data.frame that contains genotypes in rows and associated breed probabilities in
columns. This is the output of build_KBP(), build_MBP(), etc.}

\item{white_}{a logical indicating whether "white" breeds (Yorkshire and Landrace, assuming present)
should be grouped together into a white breed probability.}
}
\value{
a data.frame containing test animals in rows and breed probabilities in columns
}
\description{
Given a breed probability matrix (provided by \code{build_KBP()} or \code{build_MBP()}), and a set of test animals'
genotypes, use this function to derive breed probabilities for each test animal genotype based on
a set of chosen SNPs
}
