% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solve_composition.R
\name{solve_composition}
\alias{solve_composition}
\title{solve_composition}
\usage{
solve_composition(Y, X, names, ped = NULL, groups = NULL, mia = FALSE,
  sire = FALSE, dam = FALSE)
}
\arguments{
\item{Y}{numeric matrix of genotypes (rows) from all animals (columns) in population
coded as dosage of allele B {0, 1, 2}}

\item{X}{numeric matrix of allele frequencies from reference animals. Frequencies are
relative to allele B.}

\item{names}{character vector of names for each breed/ancestral population in the order
that they appear in X.}

\item{ped}{data.frame giving pedigree information. Must be formatted "ID", "Sire", "Dam"}

\item{groups}{list of IDs catagoriezed by breed/population. If specified, output will be a list
of results categorized by breed/population.}

\item{mia}{logical. Only applies if ped argument is supplied. If true, returns a data.frame
containing the inferred maternally inherited allele for each locus for each animal instead
of breed composition results.}

\item{sire}{logical. Only applies if ped argument is supplied. If true, returns a data.frame
containing sire genotypes for each locus for each animal instead of breed composition results.}
}
\value{
A data.frame or list of data.frames (if groups is !NULL) with breed/ancestry compostion
 results
}
\description{
Computes genome-wide breed/ancestry composition using quadratic programming on a
 batch of animals.
}

