% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_MBP.R
\name{solve_MBP}
\alias{solve_MBP}
\title{Solve MC1R-Based Breed Probabilities (MBP)}
\usage{
solve_MBP(geno, b_mat, white = FALSE)
}
\arguments{
\item{geno}{a matrix or data.frame with genotypes of test animals. SNPs in columns and 
individuals in rows. Genotypes must be coded as dosage of allele 'b' {0, 1, 2}. NOTE, only 8 SNPs
surrounding MC1R will actually be used.}

\item{b_mat}{a data.frame that contains genotypes in rows and associated breed probabilities in
columns. This is the output of build_MBP().}

\item{white}{a logical indicating whether "white" breeds (Yorkshire and Landrace, assuming present)
should be grouped together into a white breed probability. For MC1R analysis, this is set to FALSE
by default.}
}
\value{
a data.frame containing test animals in rows and KBPs in columns
}
\description{
Given a breed probability matrix (provided by \code{build_MBP()}), and a set of test animals
genotypes, use this function to derive breed probabilities for each test animal genotype based on
8 SNPs within a 2MB window around the MC1R gene.
}
