% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_KBP.R
\name{build_KBP}
\alias{build_KBP}
\title{Build reference panel for KBP}
\usage{
build_KBP(geno, map, ped = NULL, path = NULL, groups = NULL,
  haplotypes = FALSE, keep_fimpute = FALSE)
}
\arguments{
\item{geno}{a matrix or data.frame with genotypes of reference animals. SNPs in columns and 
individuals in rows. Genotypes must be coded as dosage of allele 'b' {0, 1, 2}.}

\item{map}{a data.frame containing SNP map information for each SNP present in geno}

\item{ped}{a data.frame pedigree providing family information for each individual in geno. The first
column of the pedigree is for ID, second is for sire/father ID, and third is for dam/mother ID.}

\item{path}{a character represting the path to the FImpute binary. If omitted, assumes FImpute binary
resides along PATH.}

\item{groups}{a list of character vectors with the names of IDs meant to be processed as groups. 
Names of list should be the names of the groups, with each element containing IDs.}

\item{haplotypes}{logical indicating whether build_KBP should return haplotype frequencies (TRUE)
or breed probabilities (FALSE).}

\item{keep_fimpute}{logical if TRUE, output files generated from snpTools::fimpute_run are kept. 
By default, build_KBP will use fimpute_run to build haplotypes, but as soon as haplotypes are 
extracted, fimpute output will be deleted so not to clutter hard drive space.}
}
\value{
a data.frame containing breed probabilities for each genotype made possible from the
reference animal haplotypes
}
\description{
KBP (Kit-based breed probabilities) utilize a reference panel of purebred animals' genotypes.
The genotypes are phased, and the resulting haplotypes are used to construct haplotype frequences,
genotype frequencies given breed \eqn{P(g | b)}, and finally breed probabilities given a genotype
\eqn{P(b | g)}. This function is used to construct a reference panel of breed probabilities.
}

