% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_plot.R
\name{dist_plot}
\alias{dist_plot}
\title{dist_plot}
\usage{
dist_plot(dat, breed, xlab = "Estimated Breed Composition")
}
\description{
Plotting function to inspect the distribution of results from a genome-wide
scan of breed composition. Plots the joint distribution (scatterplot) of breed composition
results of breed specified and R squared values. Marginal distributions (histograms)
on the axes.
}
\details{
@param dat data.frame of genome-wide breed composition results
 @param breed character naming the desired breed composition result in dat to plot
 @param xlab character giving label for x-axis
 @return NONE
 @export
}

