% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_composition.R
\name{sim_composition}
\alias{sim_composition}
\title{Simulate hybrid animals for breed composition}
\usage{
sim_composition(Y, X, rep = 10000, par1, par2 = NULL)
}
\arguments{
\item{Y}{numeric matrix of genotypes from all animals in population}

\item{X}{numeric matrix of allele frequencies from reference animals}

\item{rep}{integer indicating how many repititions of the simulation to perform}

\item{par1}{list of IDs catagorized by breed/population to be used as "Parent 1" in a simulated
progeny test}

\item{par2}{list of IDs catagorized by breed/population to be used as "Parent 2" in a simulated
progeny test. If null, two parents from sim1 will be chosen}
}
\description{
Use synthetic offspring to test the accuracy of breed/ancestry calculations among test population
}

