% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_plot.R
\name{sim_plot}
\alias{sim_plot}
\title{Inspect the results of sim_composition()}
\usage{
sim_plot(sim, y = "XBreed1.qp")
}
\arguments{
\item{sim}{data.frame returned from _____, containing breed composition
estimates from simulated offspring.}

\item{y}{character naming the column of sim to be plotted on y. By default,
"XBreed1.qp" refers to the estimated breed composition of breed 1.}
}
\value{
NONE
}
\description{
Generates a scatterplot to inspect the results of repeatedly simulated 
QP_SimCross output.
}

