% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_KBP.R
\name{solve_KBP}
\alias{solve_KBP}
\title{Solve KIT-Based Breed Probabilities (KBP)}
\usage{
solve_KBP(geno, b_mat, white = TRUE)
}
\arguments{
\item{geno}{a matrix or data.frame with genotypes of test animals. SNPs in columns and 
individuals in rows. Genotypes must be coded as dosage of allele 'b' {0, 1, 2}. NOTE, only 7 SNPs
surrounding KIT will actually be used.}

\item{b_mat}{a data.frame that contains genotypes in rows and associated breed probabilities in
columns. This is the output of build_KBP().}

\item{white}{a logical indicating whether "white" breeds (Yorkshire and Landrace, assuming present)
should be grouped together into a white breed probability.}
}
\value{
a data.frame containing test animals in rows and KBPs in columns
}
\description{
Given a breed probability matrix (provided by \code{build_KBP()}), and a set of test animals
genotypes, use this function to derive breed probabilities for each test animal genotype based on
7 SNPs around the KIT gene.
}
